third_party/blink/web_tests/fast/animation/scroll-animations/animation-created-with-scroll-timeline.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test instantiating an Animation with a ScrollTimeline.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
  test(t => {
    var scrollTimeline = new ScrollTimeline({ orientation: 'block' });
    var animationObject = new Animation(null, scrollTimeline);
    assert_equals(animationObject.timeline, scrollTimeline,
      "Animation timeline should be the ScrollTimeline passed in the constructor.");
  }, 'Instantiating an Animation with a ScrollTimeline should be successful.');
</script>
